direnv: Specify the path to my nix wrapper
Before this change, I would take the location of the first nix found on
the PATH and assume that was my wrapper. This doesn't work on macOS since
it doesn't add ~/.local/bin to the PATH. Instead, it would find the nix
in /nix/var/nix/profiles/default/bin. Not only would my wrapper not be
put first on the PATH, but all the programs in
/nix/var/nix/profiles/default/bin were shadowing programs in my
environment. This didn't affect me when the environment was being loaded
in fish since I add ~/.local/bin to the PATH in my fish config. It did
affect me however in VS Code, where the environment gets loaded in
the default shell, zsh.