Fix garbage collection on macOS
Unlike systemd, macOS does not have the concept of a "persistent"[1] timer so if
the computer is not on when a timer is supposed to go off, it never does. To get
around this, I run it a few times on that day, aborting later calls if an
earlier one that day succeeded.

[1]: https://www.freedesktop.org/software/systemd/man/latest/systemd.timer.html#Persistent=