Stop using sudo --validate
I did this so I would get prompted for sudo at the beginning of a rebuild
instead of halfway through it. Since then I've allowed my user to run the whole
rebuild without a password[1] so this isn't necessary. In fact, having this caused
me to get prompted for my password when I other wouldn't have.

[1]: https://github.com/bigolu/system-configurations/commit/e070b08dd12c2144e691665c14f08fdbce5bbe88