ci: Remove comment
It's not accurate since I think the reusable workflow is actually a better
choice than an action. This way the logic for making sure the task gets run once
can be centralized too. As an action, every task would have to start with a
conditional to see if it ran before.