speakers: Add an ExecStartPre to poll for network availability
Though this is service is ordered after network-online.target, this only
guarantees the network on startup, not after wake. This is because
NetworkManager shuts down before sleep and restarts after wake, but
network-online.target remains active from the moment the network is first online
until the computer shuts down. So when the system wakes from sleep, systemd sees
that network-online.target is active starts this service, but in reality, the
network is not actually online.