direnv-manual-reload: Disable adding files to the file watch
This way, I don't have to worry about nix-direnv adding to the watch list. Now
that nix-direnv can't add to the watch list, it will only reload when the files
added to the watch list by direnv-manual-reload change. Because of this, I no
longer have to use nix-direnv's manual reloading.