fix(speakerctl): Adjust suspension detection
If the network goes down less than 10 seconds after the computer resumes, the
`turn-off-speakers` script wrongly assumes that the computer is entering
susupension, instead of leaving it. This is because it greps `journalctl` for
the word "suspend". With this change, it will now look for "suspend now" which
should only be present if the computer is entering suspension.